+++ src/main/kotlin/no/nav/bidrag/dokument/arkiv/dto/DokumentoversiktFagsakQueryResponse.kt
@@ -5,7 +5,7 @@ import no.nav.bidrag.dokument.arkiv.model.JournalpostIkkeFunnetException
 data class DokumentoversiktFagsakQueryResponse(
     var data: DokumentoversiktFagsak? = null
 ) {
-    fun hentJournalpost(journalpostId: Int): Journalpost {
+    fun hentJournalpost(journalpostId: Long): Journalpost {
         return data?.dokumentoversiktFagsak?.journalposter?.find { it.journalpostId == journalpostId.toString() }
             ?: throw JournalpostIkkeFunnetException("Ingen journalpst med id: $journalpostId")
     }
